+++ src/proxy-page/proxy-page.service.ts
@@ -32,6 +32,7 @@ import fixDrawioMacro from './steps/fixDrawio';
 import fixChartMacro from './steps/fixChart';
 import fixRoadmap from './steps/fixRoadmap';
 import fixFrameAllowFullscreen from './steps/fixFrameAllowFullscreen';
+import { Content } from '../confluence/confluence.interface';
 
 @Injectable()
 export class ProxyPageService {
@@ -43,9 +44,33 @@ export class ProxyPageService {
     private jira: JiraService,
   ) {}
 
+  private initContext(
+    spaceKey: string,
+    pageId: string,
+    theme: string,
+    style: string,
+    data: Content,
+  ) {
+    this.context.Init(spaceKey, pageId, theme, style);
+    this.context.setTitle(data.title);
+    this.context.setHtmlBody(data.body.view.value);
+    this.context.setAuthor(data.history.createdBy.displayName);
+    this.context.setEmail(data.history.createdBy.email);
+    this.context.setAvatar(data.history.createdBy.profilePicture.path);
+    this.context.setWhen(data.history.createdDate);
+    if (
+      data.metadata.properties['content-appearance-published']?.value ===
+      'full-width'
+    ) {
+      this.context.setFullWidth(true);
+    } else {
+      this.context.setFullWidth(false);
+    }
+  }
+
   /**
-   * Function renderPage Service
-   *
+   * @function renderPage Service
+   * @description Render a Confluence page as read-only konviw document
    * @return Promise {string}
    * @param spaceKey {string} 'iadc' - space key where the page belongs
    * @param pageId {string} '639243960' - id of the page to retrieve
@@ -103,6 +128,7 @@ export class ProxyPageService {
 
   /**
    * @function renderSlides Service
+   * @description Render a Confluence page as konviw slide deck
    * @return Promise {string}
    * @param spaceKey {string} 'iadc' - space key where the page belongs
    * @param pageId {string} '639243960' - id of the page to retrieve
@@ -138,7 +164,7 @@ export class ProxyPageService {
   }
 
   /**
-   * getMediaCdnUrl Service
+   * @function getMediaCdnUrl Service
    * @return Promise string
    * @param uri {string} 'iadc' - URL of the media file to return
    */